# workload/ycsb: add retry handling to scan operation