--- a/pkg/workload/ycsb/ycsb.go
+++ b/pkg/workload/ycsb/ycsb.go
@@ -28,6 +28,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 	"github.com/cockroachdb/cockroach/pkg/workload"
 	"github.com/cockroachdb/cockroach/pkg/workload/histogram"
+	"github.com/lib/pq"
 	"github.com/pkg/errors"
 	"github.com/spf13/pflag"
 )
@@ -572,14 +573,45 @@ func (yw *ycsbWorker) readRow(ctx context.Context) error {
 func (yw *ycsbWorker) scanRows(ctx context.Context) error {
 	key := yw.nextReadKey()
 	scanLength := yw.scanLengthGen.Uint64()
-	res, err := yw.scanStmt.QueryContext(ctx, key, scanLength)
-	if err != nil {
-		return err
+	// We run the SELECT statement in a retry loop to handle retryable errors.
+	// The scan is large enough that it occasionally begins streaming results
+	// back to the client. If it then hits a ReadWithinUncertaintyIntervalError
+	// then it will return this error even if it is being run as an implicit
+	// transaction.
+	// TODO(nvanbenschoten): replace this with crdb.Execute when we update our
+	// vendored cockroachdb/cockroach-go library. Updating is currently running
+	// into issues because it requires us to update jackc/pgx to v4, which
+	// requires module support.
+	for {
+		err := func() error {
+			res, err := yw.scanStmt.QueryContext(ctx, key, scanLength)
+			if err != nil {
+				return err
+			}
+			defer res.Close()
+			for res.Next() {
+			}
+			return res.Err()
+		}()
+		if err == nil || !errIsRetryable(err) {
+			return err
+		}
 	}
-	defer res.Close()
-	for res.Next() {
+}
+
+func errIsRetryable(err error) bool {
+	switch t := err.(type) {
+	case *pq.Error:
+		// We look for either:
+		//  - the standard PG errcode SerializationFailureError:40001 or
+		//  - the Cockroach extension errcode RetriableError:CR000. This extension
+		//    has been removed server-side, but support for it has been left here for
+		//    now to maintain backwards compatibility.
+		return t.Code == "CR000" || t.Code == "40001"
+
+	default:
+		return false
 	}
-	return res.Err()
 }
 
 func (yw *ycsbWorker) readModifyWriteRow(ctx context.Context) error {