Fixes #46028.
Fixes #46071.

This commit adds retry handling to YCSB's "scan" operation, which is used in
workload E. We run the SELECT statement in a retry loop to handle retryable
errors. The scan is large enough that it occasionally begins streaming results
back to the client. If it then hits a ReadWithinUncertaintyIntervalError then it
will return this error even if it is being run as an implicit transaction.

We unfortunately cannot use the new crdb.Execute function (https://github.com/cockroachdb/cockroach-go/pull/57)
because updating the cockroachdb/cockroach-go library past https://github.com/cockroachdb/cockroach-go/commit/a11ba43d4f3a58987f80628602200bd6334dba64
turns out to be difficult. Doing so requires us to update jackc/pgx to v4, which
requires module support.

This change does not have a noticeable effect on workload E performance.
However, using crdb.ExecuteTx would have.

```
# before
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
   30.0s        0           3486          116.2      7.8      6.0     21.0     35.7    121.6  insert

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
   30.0s        0          66321         2210.5     14.1     11.0     37.7     71.3    192.9  scan

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
   30.0s        0          69807         2326.7     13.7     10.5     37.7     71.3    192.9


# after with implicit txn & retry loop (this commit)
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
   30.0s        0           3518          117.3      7.7      5.8     21.0     37.7    134.2  insert

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
   30.0s        0          67013         2233.7     13.9     11.0     37.7     67.1    268.4  scan

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
   30.0s        0          70531         2351.0     13.6     10.5     37.7     67.1    268.4


# after with explicit txn & retry loop  (not this commit)
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
   30.0s        0           3179          106.0      5.1      3.1     14.7     30.4    142.6  insert

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
   30.0s        0          59909         1996.9     15.7     12.1     41.9     79.7    243.3  scan

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
   30.0s        0          63088         2102.9     15.2     12.1     39.8     75.5    243.3
```

Release justification: testing only change